+++ donations/app_settings.py
@@ -1,5 +1,3 @@
-from datetime import timedelta
-
 from django.conf import settings
 
 JUST_GIVING_WEB_URL = getattr(settings, 'JUST_GIVING_WEB_URL', 'http://v3-sandbox.justgiving.com')
@@ -9,10 +7,6 @@ JUST_GIVING_APP_ID = getattr(settings, 'JUST_GIVING_APP_ID', 'XXXXX')
 
 CURRENCIES = getattr(settings, 'CURRENCIES', ['GBP', 'EUR', 'AUD', 'USD', 'ZAR', 'CAD', 'AED', 'HKD', 'SGD'])
 
-DONATION_FREQUENCIES = getattr(settings, 'DONATION_FREQUENCIES', {
-    'single': timedelta(days=0),
-})
-DONATION_PROVIDERS = getattr(settings, 'DONATION_PROVIDERS', {
-    'Just Giving': 'just_giving.SimpleDonationProvider'
-})
 VERIFY_API_URL_NAME = getattr(settings, 'DONATION_VERIFY_API_URL_NAME', 'donations:api:verify')
+
+VERIFY_FINISHED_URL = getattr(settings, 'DONATION_VERIFY_FINISHED_URL', '/donate/thanks/')