+++ donations/models.py
@@ -1,10 +1,11 @@
 import logging
-from importlib import import_module
 
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
 from djmoney.models.fields import MoneyField
+
+from .providers.just_giving import SimpleDonationProvider as JustGivingProvider
 
 logger = logging.getLogger(__name__)
 
@@ -21,44 +22,42 @@ class Frequency(models.Model):
         return "{0} ({1})".format(self.name, self.interval)
 
 
-class DonationProvider(models.Model):
-    """External provider that handle the donations"""
-
-    name = models.CharField(max_length=100, default='BlankProvider')
-    description = models.CharField(max_length=255, default='')
-    klass = models.CharField(max_length=255, default='')
-
-    def get_provider_class(self):
-        """get the class of the donation provider.
-        I have hardcoded the first bit to prevent any module being imported"""
-        module_name, klass_name = self.klass.rsplit('.', 1)
-        module = import_module('donations.providers.{0}'.format(module_name))
-        return getattr(module, klass_name)
-
-    def __str__(self):
-        return self.name
-
-
-DONATION_SOURCE = ["DirectDonations", "SponsorshipDonations", "Ipdd", "Sms"]
-DONATION_SOURCE = [(i, i) for i in DONATION_SOURCE]
-
-DONATION_STATUSES = ["Accepted", "Rejected", "Cancelled", "Refunded", "Pending", "Unverified"]
-DONATION_STATUSES = [(i, i) for i in DONATION_STATUSES]
-
-
 class Donation(models.Model):
     """(Abstract representation of a Donation)"""
+    SOURCE_CHOICES = [
+        ("DirectDonations", "Direct Donations"),
+        ("SponsorshipDonations", "Sponsorship Donations"),
+        ("Ipdd", "IPDD"),
+        ("Sms", "SMS"),
+    ]
+
+    class Statuses:
+        ACCEPTED = "Accepted"
+        REJECTED = "Rejected"
+        CANCELLED = "Cancelled"
+        REFUNDED = "Refunded"
+        PENDING = "Pending"
+        UNVERIFIED = "Unverified"
+        CHOICES = [
+            (ACCEPTED, "Accepted"),
+            (REJECTED, "Rejected"),
+            (CANCELLED, "Cancelled"),
+            (REFUNDED, "Refunded"),
+            (PENDING, "Pending"),
+            (UNVERIFIED, "Unverified"),
+        ]
 
     # https://github.com/django-money/django-money
     amount = MoneyField(max_digits=10, decimal_places=2, default_currency='GBP')
-    provider = models.ForeignKey(DonationProvider, related_name='donations', on_delete=models.CASCADE)
     frequency = models.ForeignKey(Frequency, related_name='donations', on_delete=models.CASCADE)
     datetime = models.DateTimeField(default=timezone.now)
     donor = models.ForeignKey(settings.AUTH_USER_MODEL, blank=True, null=True, related_name='donations',
                               on_delete=models.CASCADE)
-    donor_display_name = models.CharField(blank=True, null=True, max_length=255,
-                                          help_text="display name returned by the provider")
-    status = models.CharField(default='Unverified', choices=DONATION_STATUSES, max_length=50)
+    donor_display_name = models.CharField(
+        blank=True, null=True, max_length=255,
+        help_text="display name returned by the provider",
+    )
+    status = models.CharField(default=Statuses.UNVERIFIED, choices=Statuses.CHOICES, max_length=50)
     is_verified = models.BooleanField(default=False)
     finished_uri = models.URLField(blank=True)
 
@@ -66,8 +65,10 @@ class Donation(models.Model):
     local_amount = MoneyField(max_digits=10, decimal_places=2, default_currency='GBP')
     message = models.CharField(blank=True, null=True, max_length=255)
     est_tax_reclaim = models.DecimalField(blank=True, null=True, max_digits=10, decimal_places=2)
-    provider_source = models.CharField(blank=True, null=True, max_length=50, choices=DONATION_SOURCE,
-                                       help_text="source of the donation from within a provider")
+    provider_source = models.CharField(
+        blank=True, null=True, max_length=50, choices=SOURCE_CHOICES,
+        help_text="source of the donation from within a provider"
+    )
 
     def __str__(self):
         if self.local_amount:
@@ -75,7 +76,8 @@ class Donation(models.Model):
         return "{0} at {1}".format(self.amount, self.datetime.strftime('%Y/%m/%d %H:%M:%S'))
 
     def get_provider(self):
-        return self.provider.get_provider_class()(self)
+        """Hook to enable retrieving another provider"""
+        return JustGivingProvider(self)
 
     def get_value(self):
         return self.amount.amount