+++ HISTORY.rst
@@ -3,6 +3,10 @@
 History
 -------
 
+* Removed the ``DonationProvider`` model, there is only 1 possible
+* Removed the ``DONATION_FREQUENCIES`` settings and magic to auto-create frequencies
+* Added a method to handle monthly donations with Just Giving
+
 1.1.0 (2018-05-31)
 ++++++++++++++++++